+++ beacon-chain/core/state/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/state",
     visibility = [
         "//beacon-chain:__subpackages__",
+        "//shared/interop:__pkg__",
         "//tools/genesis-state-gen:__pkg__",
     ],
     deps = [
