+++ shared/interop/BUILD.bazel
@@ -2,12 +2,21 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["generate_keys.go"],
+    srcs = [
+        "generate_genesis_state.go",
+        "generate_keys.go",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/shared/interop",
     visibility = ["//visibility:public"],
     deps = [
+        "//beacon-chain/core/state:go_default_library",
+        "//proto/beacon/p2p/v1:go_default_library",
+        "//proto/eth/v1alpha1:go_default_library",
         "//shared/bls:go_default_library",
         "//shared/hashutil:go_default_library",
+        "//shared/params:go_default_library",
+        "//shared/trieutil:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
+        "@com_github_prysmaticlabs_go_ssz//:go_default_library",
     ],
 )
