+++ tools/genesis-state-gen/BUILD.bazel
@@ -6,15 +6,9 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/tools/genesis-state-gen",
     visibility = ["//visibility:private"],
     deps = [
-        "//beacon-chain/core/state:go_default_library",
-        "//proto/eth/v1alpha1:go_default_library",
-        "//shared/bls:go_default_library",
-        "//shared/hashutil:go_default_library",
         "//shared/interop:go_default_library",
         "//shared/params:go_default_library",
-        "//shared/trieutil:go_default_library",
         "@com_github_ghodss_yaml//:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_go_ssz//:go_default_library",
     ],
 )
